requirements.txt
aws-cdk-lib==2.171.1
constructs>=10.0.0,<11.0.0
psycopg2-binary==2.9.9
pyyaml==6.0.2
boto3==1.35.87
botocore==1.35.87